separated MAC evaluation into separate file
