+++ include/cstone/tree/octree_essential.hpp
@@ -43,141 +43,13 @@
 
 #include <vector>
 
-#include "cstone/cuda/annotation.hpp"
-#include "cstone/sfc/morton.hpp"
 #include "cstone/halos/boxoverlap.hpp"
+#include "macs.hpp"
 #include "octree_internal.hpp"
 #include "traversal.hpp"
 
 namespace cstone
 {
-
-CUDA_HOST_DEVICE_FUN
-template<int Period>
-constexpr int rangeSeparation(int a, int b, int c, int d, bool pbc)
-{
-    assert(a < b && c < d);
-    int cb = c - b;
-    int ad = a - d;
-    int cbPbc = (pbc) ? pbcDistance<Period>(cb) : cb;
-    int adPbc = (pbc) ? pbcDistance<Period>(ad) : ad;
-    return (cb >= 0 || ad >= 0) * stl::min(stl::abs(cbPbc), stl::abs(adPbc));
-}
-
-/*! @brief return the smallest distance squared between two points on the surface of the AABBs @p a and @p b
- *
- * @tparam T     float or double
- * @tparam I     32- or 64-bit unsigned integer
- * @param a      a box, specified with integer coordinates in [0:2^21]
- * @param b
- * @param box    floating point coordinate bounding box
- * @return       the square of the smallest distance between a and b
- */
-template<class T, class I>
-CUDA_HOST_DEVICE_FUN
-T minDistanceSq(IBox a, IBox b, const Box<T>& box)
-{
-    constexpr size_t maxCoord = 1u<<maxTreeLevel<I>{};
-    constexpr T unitLengthSq  = T(1.) / (maxCoord * maxCoord);
-
-    size_t dx = rangeSeparation<maxCoord>(a.xmin(), a.xmax(), b.xmin(), b.xmax(), box.pbcX());
-    size_t dy = rangeSeparation<maxCoord>(a.ymin(), a.ymax(), b.ymin(), b.ymax(), box.pbcY());
-    size_t dz = rangeSeparation<maxCoord>(a.zmin(), a.zmax(), b.zmin(), b.zmax(), box.pbcZ());
-    // the maximum for any integer is 2^21-1, so we can safely square each of them
-    return ((dx*dx)*box.lx()*box.lx() + (dy*dy)*box.ly()*box.ly() +
-            (dz*dz)*box.lz()*box.lz()) * unitLengthSq;
-}
-
-//! @brief return longest edge length of box @p b
-template<class T, class I>
-CUDA_HOST_DEVICE_FUN
-T nodeLength(IBox b, const Box<T>& box)
-{
-    constexpr int maxCoord = 1u<<maxTreeLevel<I>{};
-    constexpr T unitLength = T(1.) / maxCoord;
-
-    // IBoxes for octree nodes are assumed cubic, only box can be rectangular
-    return (b.xmax() - b.xmin()) * unitLength * box.maxExtent();
-}
-
-/*! @brief evaluate minimum distance MAC
- *
- * @param a            target cell
- * @param b            source cell
- * @param box          coordinate bounding box
- * @param invThetaSq   inverse theta squared
- * @return             true if MAC fulfilled, false otherwise
- *
- * Note: Mac is valid for any point in a w.r.t to box b, therefore only the
- * size of b is relevant.
- */
-template<class T, class I>
-CUDA_HOST_DEVICE_FUN
-bool minDistanceMac(IBox a, IBox b, const Box<T>& box, float invThetaSq)
-{
-    T dsq = minDistanceSq<T, I>(a, b, box);
-    // equivalent to "d > l / theta"
-    T bLength = nodeLength<T, I>(b, box);
-    return dsq > bLength * bLength * invThetaSq;
-}
-
-template<class T, class I>
-CUDA_HOST_DEVICE_FUN
-void markMacPerBox(IBox target, const Octree<I>& octree, const Box<T>& box,
-                   float invThetaSq, I focusStart, I focusEnd, char* markings)
-{
-    auto checkAndMarkMac = [target, &octree, &box, invThetaSq, focusStart, focusEnd, markings](TreeNodeIndex idx)
-    {
-        I nodeStart = octree.codeStart(idx);
-        I nodeEnd   = octree.codeEnd(idx);
-        // if the tree node with index idx is fully contained in the focus, we stop traversal
-        if (containedIn(nodeStart, nodeEnd, focusStart, focusEnd)) { return false; }
-
-        IBox sourceBox = makeIBox(nodeStart, nodeEnd);
-
-        bool violatesMac = !minDistanceMac<T, I>(target, sourceBox, box, invThetaSq);
-        if (violatesMac) { markings[idx] = 1; }
-
-        return violatesMac;
-    };
-
-    traverse(octree, checkAndMarkMac, [](TreeNodeIndex){});
-}
-
-/*! @brief Mark all leaf nodes that fail the MAC paired with leaf nodes from a given range
- *
- * @tparam T                float or double
- * @tparam I                32- or 64-bit unsigned integer
- * @param[in]  octree       octree, including internal part
- * @param[in]  box          global coordinate bounding box
- * @param[in]  firstLeaf    first leaf index of the cornerstone tree used to build @p octree
- *                          to check for nodes failing the minimum distance Mac
- * @param[in]  lastLeaf     last leaf index
- * @param[in]  invThetaSq   1./theta^2
- * @param[out] markings     array of length @p octree.nTreeNodes(), each position i
- *                          will be set to 1, if node with index i fails the MAC paired with any
- *                          of the leaf nodes with leaf index in [firstLeaf:lastLeaf]
- */
-template<class T, class I>
-void markMac(const Octree<I>& octree, const Box<T>& box, I focusStart, I focusEnd,
-             float invThetaSq, char* markings)
-
-{
-    std::fill(markings, markings + octree.nTreeNodes(), 0);
-
-    // find the minimum possible number of octree node boxes to cover the entire focus
-    TreeNodeIndex numFocusBoxes = spanSfcRange(focusStart, focusEnd);
-    std::vector<I> focusCodes(numFocusBoxes + 1);
-    spanSfcRange(focusStart, focusEnd, focusCodes.data());
-    focusCodes.back() = focusEnd;
-
-    #pragma omp parallel for schedule(static)
-    for (TreeNodeIndex i = 0; i < numFocusBoxes; ++i)
-    {
-        IBox target = makeIBox(focusCodes[i], focusCodes[i+1]);
-        markMacPerBox(target, octree, box, invThetaSq, focusStart, focusEnd, markings);
-    }
-}
 
 template<class I>
 inline CUDA_HOST_DEVICE_FUN
